Add upsert_program_ex function (#41)
This does the same things as upsert_program, but defines all arguments
as function parameters, which means they can be found easily, can be
documented properly, can be given in any order and do no longer need to
be pre-declared.

Added the function in addition to the existing function with an '_ex'
suffix. Suggest renaming it to the without-'_ex' name with the next
major release (5.0).

Add tests for upsert_program_ex and rename some parameters for consistency between the updating function and the ProcessingProgram model object.