Defer loading of externalId columns
These columns result in a UnicodeDecode error when querying the DLS tables with a
mysql+mysqlconnector connection. Using mysql+pymysql there is no such error.
Setting the columns to deferred at least means that the tables can be queried
without error, and I think it is unlikely that these columns will be widely used
via the ipsyb-api.
See also:
https://stackoverflow.com/questions/54182365/i-get-an-error-when-get-my-binary-data-from-mysql-db-using-sqlalchemy