Change package to /src layout (#133)
This moves all source files into /src/ispyb and thus resolves #131.
This source layout provides a clean package boundary and prevents importing
non-package files such as setup.py and the /bin directory.

setup.py / setup.cfg were updated accordingly to support the packaging change,
and most package metadata has now moved into the .cfg file. 

The package used to be described as 'beta', but we have been treating it as production
for a long time now, so we might as well change that descriptor to 'stable'.