Add a debug facility to the sqlalchemy interface (#136)
When ispyb.sqlalchemy.enable_debug_logging() is called all future SQL
commands executed via the sqlalchemy interface will be logged through
the Python logging framework. This should not be used in production
as it can be expensive, can leak sensitive information, and, once
enabled, cannot be disabled.

Added a '--debug' flag to the 'ispyb.last_data_collections_on'
command to demonstrate. Fix up the sqlalchemy logic in
the command to reduce number of needed SQL queries
from 3n+1 to 1.